App/app/requirements.txt
requests==2.32.3
python-dotenv==1.1.0
streamlit==1.44.1
pillow==10.4.0